# api/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Backend for the Birga caregiver app.",
        "title": "Birga API",
        "contact": {},
        "version": "v1"
    },
    "basePath": "/",
    "paths": {
...
        "/v1/children/{id}/completions": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "children"
                ],
                "summary": "lists a child's completions, newest first",
                "parameters": [
                    {
                        "type": "string",
                        "description": "child id (UUID)",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "page size (default 20, max 100)",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "offset",
                        "name": "offset",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/rest.R"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/rest.completionListView"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/rest.BadRequestResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/rest.UnauthorizedResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/rest.NotFoundResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/rest.InternalServerErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "- only the child's parents can record (404 otherwise); the activity must be published (404)\n- the day is the current date in Uzbekistan (UTC+5); repeating the same activity on the same\nday returns the existing completion, replacing its note if a new one is sent\n- note is the caregiver's optional reflection, at most 1000 characters (422)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "children"
                ],
                "summary": "marks an activity done for a child today",
                "parameters": [
                    {
                        "type": "string",
                        "description": "child id (UUID)",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "completion",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/rest.CompleteActivityRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/rest.R"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/rest.completionView"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/rest.BadRequestResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/rest.UnauthorizedResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/rest.NotFoundResponse"
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "schema": {
                            "$ref": "#/definitions/rest.UnprocessableContentResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/rest.InternalServerErrorResponse"
                        }
                    }
                }
            }
        },
        "/v1/children/{id}/recommendation": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "- only the child's parents can ask (404 otherwise); the child's age is clamped to 2..6\n- goal and minutes narrow the choice; minutes keeps activities with duration_minutes \u003c= minutes\n- activities the child has never done come first, then the one done longest ago;\nties are broken per child and per day, so the pick stays the same all day until it is done\n- 404 when no published activity matches",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "children"
                ],
                "summary": "today's recommended activity for a child",
                "parameters": [
                    {
                        "type": "string",
                        "description": "child id (UUID)",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "enum": [
                            "language",
                            "motor",
                            "cognitive",
                            "social",
                            "emotional"
                        ],
                        "type": "string",
                        "description": "development goal",
                        "name": "goal",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "time available, in minutes",
                        "name": "minutes",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/rest.R"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/rest.activityView"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/rest.BadRequestResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/rest.UnauthorizedResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/rest.NotFoundResponse"
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "schema": {
                            "$ref": "#/definitions/rest.UnprocessableContentResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/rest.InternalServerErrorResponse"
                        }
                    }
                }
            }
        },
        "/v1/children/{id}/streak": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "- current: consecutive days with a completion up to today, or up to yesterday while\ntoday has none yet (the streak is not lost until a whole day is missed)\n- longest: the longest such run; this_week: days with a completion since Monday\n- days are counted in Uzbekistan time (UTC+5)",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "children"
                ],
                "summary": "a child's streak and progress",
                "parameters": [
                    {
                        "type": "string",
                        "description": "child id (UUID)",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/rest.R"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/rest.streakView"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/rest.BadRequestResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/rest.UnauthorizedResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/rest.NotFoundResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/rest.InternalServerErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "rest.BadRequestResponse": {
            "type": "object",
            "properties": {
                "data": {},
                "error_code": {
                    "allOf": [
                        {
                            "$ref": "#/definitions/rest.errCode"
                        }
                    ],
                    "example": -11
                },
                "error_note": {
                    "type": "string",
                    "example": "limit must be a positive integer"
                },
                "status": {
                    "allOf": [
                        {
                            "$ref": "#/definitions/rest.statusType"
                        }
                    ],
                    "example": "Failure"
                }
            }
        },
        "rest.CompleteActivityRequest": {
            "type": "object",
            "properties": {
                "activity_id": {
                    "type": "string",
                    "example": "7b0c1f1e-2d7a-4d8e-9a55-0f4a0d7f9c11"
                },
                "note": {
                    "description": "caregiver's reflection, optional",
                    "type": "string",
                    "example": "Qizil rangni birinchi topdi"
                }
            }
        },
...
        "rest.InternalServerErrorResponse": {
            "type": "object",
            "properties": {
                "data": {},
                "error_code": {
                    "allOf": [
                        {
                            "$ref": "#/definitions/rest.errCode"
                        }
                    ],
                    "example": -50
                },
                "error_note": {
                    "type": "string",
                    "example": "internal error"
                },
                "status": {
                    "allOf": [
                        {
                            "$ref": "#/definitions/rest.statusType"
                        }
                    ],
                    "example": "Failure"
                }
            }
        },
        "rest.NotFoundResponse": {
            "type": "object",
            "properties": {
                "data": {},
                "error_code": {
                    "allOf": [
                        {
                            "$ref": "#/definitions/rest.errCode"
                        }
                    ],
                    "example": -30
                },
                "error_note": {
                    "type": "string",
                    "example": "activity not found"
                },
                "status": {
                    "allOf": [
                        {
                            "$ref": "#/definitions/rest.statusType"
                        }
                    ],
                    "example": "Failure"
                }
            }
        },
        "rest.R": {
            "type": "object",
            "properties": {
                "data": {},
                "error_code": {
                    "$ref": "#/definitions/rest.errCode"
                },
                "error_note": {
                    "type": "string"
                },
                "status": {
                    "$ref": "#/definitions/rest.statusType"
                }
            }
        },
...
        "rest.UnauthorizedResponse": {
            "type": "object",
            "properties": {
                "data": {},
                "error_code": {
                    "allOf": [
                        {
                            "$ref": "#/definitions/rest.errCode"
                        }
                    ],
                    "example": -20
                },
                "error_note": {
                    "type": "string",
                    "example": "unauthorized"
                },
                "status": {
                    "allOf": [
                        {
                            "$ref": "#/definitions/rest.statusType"
                        }
                    ],
                    "example": "Failure"
                }
            }
        },
        "rest.UnprocessableContentResponse": {
            "type": "object",
            "properties": {
                "data": {},
                "error_code": {
                    "allOf": [
                        {
                            "$ref": "#/definitions/rest.errCode"
                        }
                    ],
                    "example": -10
                },
                "error_note": {
                    "type": "string",
                    "example": "unknown goal \"flying\""
                },
                "status": {
                    "allOf": [
                        {
                            "$ref": "#/definitions/rest.statusType"
                        }
                    ],
                    "example": "Failure"
                }
            }
        },
...
        "rest.activityView": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "description_ru": {
                    "type": "string"
                },
                "description_uz": {
                    "type": "string"
                },
                "duration_minutes": {
                    "type": "integer",
                    "example": 10
                },
                "goal": {
                    "type": "string",
                    "example": "cognitive"
                },
                "id": {
                    "type": "string",
                    "example": "7b0c1f1e-2d7a-4d8e-9a55-0f4a0d7f9c11"
                },
                "is_published": {
                    "type": "boolean",
                    "example": true
                },
                "max_age": {
                    "type": "integer",
                    "example": 5
                },
                "min_age": {
                    "type": "integer",
                    "example": 3
                },
                "title_ru": {
                    "type": "string",
                    "example": "Цветные камни"
                },
                "title_uz": {
                    "type": "string",
                    "example": "Rangli toshlar"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "rest.completionListView": {
            "type": "object",
            "properties": {
                "items": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/rest.completionView"
                    }
                },
                "limit": {
                    "type": "integer",
                    "example": 20
                },
                "offset": {
                    "type": "integer",
                    "example": 0
                },
                "total": {
                    "type": "integer",
                    "example": 12
                }
            }
        },
        "rest.completionView": {
            "type": "object",
            "properties": {
                "activity_id": {
                    "type": "string",
                    "example": "7b0c1f1e-2d7a-4d8e-9a55-0f4a0d7f9c11"
                },
                "child_id": {
                    "type": "string",
                    "example": "3f1d2c4b-8a9e-4b7c-9d2e-1a2b3c4d5e6f"
                },
                "completed_on": {
                    "type": "string",
                    "example": "2026-10-03"
                },
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "string",
                    "example": "5d2f8a90-1c3b-4e6f-8a7d-9b0c1d2e3f4a"
                },
                "note": {
                    "type": "string",
                    "example": "Qizil rangni birinchi topdi"
                },
                "user_id": {
                    "type": "string",
                    "example": "9a8b7c6d-5e4f-4a3b-2c1d-0e9f8a7b6c5d"
                }
            }
        },
        "rest.errCode": {
            "type": "integer",
            "enum": [
                0,
                -10,
                -11,
                -20,
                -21,
                -30,
                -40,
                -50,
                -60,
                -70
            ],
            "x-enum-comments": {
                "_errCodeBadRequest": "for http 400",
                "_errCodeConflict": "for http 409",
                "_errCodeForbidden": "for http 403",
                "_errCodeInternalErr": "for http 500",
                "_errCodeNoError": "Success code.",
                "_errCodeNotFound": "for http 404",
                "_errCodeRateLimited": "for http 429",
                "_errCodeUnauthorized": "for http 401",
                "_errCodeUnavailable": "for http 503",
                "_errCodeValidation": "for http 422"
            },
            "x-enum-descriptions": [
                "Success code.",
                "for http 422",
                "for http 400",
                "for http 401",
                "for http 403",
                "for http 404",
                "for http 409",
                "for http 500",
                "for http 503",
                "for http 429"
            ],
            "x-enum-varnames": [
                "_errCodeNoError",
                "_errCodeValidation",
                "_errCodeBadRequest",
                "_errCodeUnauthorized",
                "_errCodeForbidden",
                "_errCodeNotFound",
                "_errCodeConflict",
                "_errCodeInternalErr",
                "_errCodeUnavailable",
                "_errCodeRateLimited"
            ]
        },
...
        "rest.statusType": {
            "type": "string",
            "enum": [
                "Success",
                "Failure"
            ],
            "x-enum-varnames": [
                "_statusSuccess",
                "_statusFailure"
            ]
        },
        "rest.streakView": {
            "type": "object",
            "properties": {
                "completed_today": {
                    "type": "boolean",
                    "example": true
                },
                "current": {
                    "type": "integer",
                    "example": 3
                },
                "last_completed_on": {
                    "type": "string",
                    "example": "2026-10-03"
                },
                "longest": {
                    "type": "integer",
                    "example": 7
                },
                "this_week": {
                    "type": "integer",
                    "example": 2
                },
                "total": {
                    "type": "integer",
                    "example": 12
                }
            }
        },
...
    },
    "securityDefinitions": {
        "AdminKey": {
            "type": "apiKey",
            "name": "X-Admin-Key",
            "in": "header"
        },
        "BearerAuth": {
            "description": "Access token from /v1/auth/signup or /v1/auth/refresh, as \"Bearer \u003ctoken\u003e\".",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}